Created a single component for the add components
